chore: drop 22 unused deps, share the re-key generator, refresh account docs
Three independent cleanups from the maintainability audit, grouped because none of
them changes behaviour and each is reviewable on its own.

**22 unused dependencies, across 9 crates.** Every one verified the same way: the
entry deleted from its manifest, then `cargo check -p <pkg> --all-targets`
recompiled clean. Only `[dependencies]` was touched, never `[dev-dependencies]`.

The payoff is larger than the count suggests, because unused deps drag their own
graphs behind them. Eleven packages leave the lockfile entirely, and three of them
were **duplicate major versions** of crates the build genuinely uses:

    axum       0.7.9 + 0.8.8   ->  0.7.9      (unused axum-extra 0.10 wanted axum 0.8)
    sha2       0.9.9 + 0.10.9  ->  0.10.9     (unused ed25519-consensus wanted sha2 0.9)
    dashmap    5.5.3 + 6.1.0   ->  6.1.0      (mero-auth pinned 5.5.3, never used it)

One unused entry in `merod` was pulling five packages: `ed25519-consensus`,
`sha2 0.9`, `digest 0.9`, `curve25519-dalek-ng`, `subtle-ng`. `calimero-context`
declared `libp2p` and never referenced it; `mero-auth` declared six, including the
workspace crates `calimero-primitives` and `calimero-store`, so four workspace
edges disappear from the rebuild fan-out too.

**`generate_struct_rekey` had two copies.** `#[app::state]` and
`#[derive(Mergeable)]` each carried one, identical apart from whether `Fields` was
imported or written `syn::Fields`. That is a bad place for two copies to live:
re-keying is where #2577 (*custom struct CRDT values lost data*) and #2581
(*nested-through-custom-struct silently LWW'd*) both were, and both were silent data
loss rather than a failure. A root struct's cascade drifting from a nested one's is
that shape of bug again, so the cascade now lives once in `macros/src/rekey.rs`
with that reasoning beside it.

**`crates/account/AGENTS.md` documented the pre-#3566 API.** Seven symbols that no
longer exist, none of the four that replaced them — an agent or engineer following
it wrote code that would not compile. My omission: #3566 touched no AGENTS.md, and
#3571 updated only authz's. Rewritten: the API table, both ASCII flow diagrams, the
module map (now including `signed.rs`), the Key Files table, and the invariants,
plus new ones for `Verified<T>` being unforgeable, `AccountMemberEndorsement`
deliberately not implementing `RootSigned`, and `AccountProof<T>` being
borsh-identical to the fields it replaced. Checked mechanically: all 70 identifiers
the file claims exist in `crates/account/src`, and no stale symbol remains.

Also demotes `verify_root_signed` to `pub(crate)`. It has no caller outside the
crate, and an unused `pub fn` in a repo that forbids dead code is a claim about an
API nobody asked for. It is mine, added in #3566.

`crates/sdk/AGENTS.md` gains the new module. No public API, wire format or
behaviour changes.

4947 workspace tests pass — including `calimero-account` (60), `calimero-sdk-macros`
(85) and both trybuild suites, `tests/macros.rs` and `tests/compile_fail.rs`, which
are what would catch a broken derive. clippy `-D warnings` and `cargo fmt` clean.
