+++ crates/auth/Cargo.toml
@@ -12,17 +12,11 @@ publish = true
 aes-gcm = "0.10.3"
 async-trait.workspace = true
 axum = { workspace = true, features = ["macros"] }
-axum-extra = { version = "0.10.1", features = ["cookie", "cookie-signed", "cookie-private"] }
 base64.workspace = true
-borsh.workspace = true
-bs58.workspace = true
-calimero-primitives.workspace = true
-calimero-store.workspace = true
 chrono = { workspace = true, features = ["serde"] }
 clap = { workspace = true, features = ["derive", "env"] }
 config = { version = "0.13", default-features = false, features = ["json", "toml"] }
 ctor = "0.1"
-dashmap = "5.5.3"
 eyre.workspace = true
 hex.workspace = true
 jsonwebtoken.workspace = true
