+++ crates/context/Cargo.toml
@@ -18,8 +18,6 @@ eyre.workspace = true
 semver.workspace = true
 futures-util.workspace = true
 hex.workspace = true
-libp2p.workspace = true
-memchr.workspace = true
 ouroboros.workspace = true
 prometheus-client.workspace = true
 rand.workspace = true
