+++ crates/context/primitives/Cargo.toml
@@ -12,9 +12,7 @@ publish = true
 calimero-account.workspace = true
 actix.workspace = true
 async-stream.workspace = true
-blake3.workspace = true
 borsh = { workspace = true, features = ["derive"] }
-ed25519-dalek.workspace = true
 eyre.workspace = true
 futures-util.workspace = true
 hex.workspace = true
