+++ crates/dag/Cargo.toml
@@ -11,7 +11,6 @@ publish = true
 [dependencies]
 async-trait.workspace = true
 borsh = { workspace = true, features = ["derive"] }
-eyre.workspace = true
 serde = { workspace = true, features = ["derive"] }
 thiserror.workspace = true
 calimero-storage.workspace = true
