+++ crates/governance-store/Cargo.toml
@@ -12,14 +12,11 @@ publish = true
 actix.workspace = true
 async-trait.workspace = true
 borsh.workspace = true
-bs58.workspace = true
 eyre.workspace = true
 hex.workspace = true
 libp2p.workspace = true
-ouroboros.workspace = true
 prometheus-client.workspace = true
 rand.workspace = true
-serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
