+++ crates/node/primitives/Cargo.toml
@@ -43,9 +43,7 @@ flate2.workspace = true
 hex.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
-serde_json_canonicalizer.workspace = true
 ed25519-dalek.workspace = true
-bs58.workspace = true
 sha2.workspace = true
 tar.workspace = true
 base64.workspace = true
