+++ crates/op-adapter/Cargo.toml
@@ -11,7 +11,6 @@ publish = false
 [dependencies]
 calimero-context-config.workspace = true
 calimero-governance-types.workspace = true
-sha2.workspace = true
 
 calimero-account.workspace = true
 calimero-op.workspace = true
